update source to current version in readme
I originally copy/pasted example out of this readme. It seemed to work, but events wouldn't fire. Eventually I used a copy of the dist itself, then guessed that filename was wrong, and indeed that was true.

Also added a note about the (to me) confusing way the path works for the file.